requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
python-dotenv>=1.0.0
slack-sdk>=3.21.0
boto3>=1.28.0
